+++ CHANGELOG.md
@@ -1,3 +1,11 @@
+## 0.2.0
+
+**Breaking**
+* Removed "noSelectionText" parameter
+
+**Changes**
+* Removed selected timeframe text as it has no direct correlation to the timeframe selector
+
 ## 0.1.0
 
 **Breaking**